test: fold textunmarshaler tests into unmarshal_test.go
#137 added unmarshal_textunmarshaler_test.go after this branch's merge-base,
so the consolidation never folded it in. Rebasing on main surfaced it as a
leftover topical, per-issue test file — violating the one-file-per-source
rule this PR establishes. Move its TestUnmarshal*/Example funcs into the
unmarshal.go sibling (unmarshal_test.go) and delete the topical file.

Refs #138
